refactor(featureFlag): Use feature flags in every route, not catch-all
