+++ app/routes/prozesskostenhilfe.antrag._index.tsx
@@ -1,9 +1,9 @@
 import { type LoaderFunctionArgs } from "@remix-run/node";
-import { throw404OnProduction } from "~/services/errorPages/throw404";
+import { throw404IfFeatureFlagEnabled } from "~/services/errorPages/throw404";
 import { loader as lastStepLoader } from "~/services/flow/server/lastStep";
 export { FormFlowPage as default } from "~/routes/shared/components/FormFlowPage";
 
-export function loader(args: LoaderFunctionArgs) {
-  throw404OnProduction();
+export async function loader(args: LoaderFunctionArgs) {
+  await throw404IfFeatureFlagEnabled("showProzesskostenhilfeFlow");
   return lastStepLoader(args);
 }
