+++ app/routes/prozesskostenhilfe.tsx
@@ -4,7 +4,7 @@ import { fetchMeta } from "~/services/cms/index.server";
 import { throw404IfFeatureFlagEnabled } from "~/services/errorPages/throw404";
 
 export async function loader({ request }: LoaderFunctionArgs) {
-  await throw404IfFeatureFlagEnabled(request);
+  await throw404IfFeatureFlagEnabled("showProzesskostenhilfeFlow");
   const { pathname } = new URL(request.url);
   const filterValue = `/${pathname.split("/").at(1) ?? ""}`;
   return json({ meta: await fetchMeta({ filterValue }) });
